Arrangement classes moved to legacy folder
